microservices/orion-news-fetcher/requirements.txt
google-cloud-firestore
google-cloud-pubsub
requests
finnhub-python
functions-framework
google-cloud-bigquery
pandas